refactor(cli): scope resolution and projected loading leave apply.go
`apply.go` was the accidental home of two clusters every scope-aware command
depends on. Scope resolution — `resolveScope`/`resolveScopeFlags`, the project
discovery walk, the interactive scope prompt and `--no-input`/TTY probes — is
consumed by `status`, `diff`, `check`, `reconcile`, `import`, `explain`,
`migrate subagents` and the component listers; `apply` is one caller among a
dozen. The same is true of `loadProjectedForScope` and the plugin-projection
overlay it owns.

They move to `scope.go` and `load.go`. `scope.go` sits beside `scope_flags.go`
(which declares the root flags and the per-command stance) without merging into
it: one file answers "may this command see --scope", the other "what does
--scope resolve to here".

Pure relocation inside package cli. `runApplyPipeline` stays in `apply.go` and
still calls `loadProjectedForScope` before `state.Load`, so the ordering guard
in apply_state_order_internal_test.go — which reads apply.go by path — keeps
reading the function it guards.
